refactor(features): Move struct to flox-rust-sdk
So that we don't duplicate the `Features` struct, or subsets of it, in
two places. This does have a couple of unfortunate side-effects though:

1. We have to thread the values through to more function calls because
   we can no longer call parse on demand.

2. The `features` field is now available from both `Config` and `Flox`.
   It would be nice if it was only available from `Flox` but we can't
   make it private from `Config` because it's used and set by a bunch of
   places.

I looked at pulling `Features.services` out to it's own type at the same
time, to make it easier to change the default later, but whilst the
default is `false` it violates a clippy rule, so it doesn't seem worth
doing at the moment:

- https://github.com/flox/flox/pull/1712#discussion_r1666985785
- https://rust-lang.github.io/rust-clippy/master/index.html#/derivable_impls
