+++ cli/flox/src/utils/search.rs
@@ -19,7 +19,6 @@ use flox_rust_sdk::models::search::{
 use log::debug;
 
 use crate::commands::{detect_environment, UninitializedEnvironment};
-use crate::config::features::Features;
 
 pub const SEARCH_INPUT_SEPARATOR: &'_ str = ":";
 pub const DEFAULT_DESCRIPTION: &'_ str = "<no description provided>";
@@ -88,13 +87,9 @@ pub(crate) fn construct_search_params(
     manifest: Option<PathOrJson>,
     global_manifest: PathOrJson,
     lockfile: PathOrJson,
+    search_strategy: flox_rust_sdk::models::search::SearchStrategy,
 ) -> Result<SearchParams> {
-    let query = Query::new(
-        search_term,
-        Features::parse()?.search_strategy,
-        results_limit,
-        true,
-    )?;
+    let query = Query::new(search_term, search_strategy, results_limit, true)?;
     let params = SearchParams {
         manifest,
         global_manifest,
